requirements.txt
certifi==2019.9.11
cffi==1.13.1
cryptography==2.8
Django==2.2.6
future==0.18.2
image==1.5.27
Pillow==6.2.1
pycparser==2.19
pytesseract==0.3.0
python-telegram-bot==12.2.0
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
tornado==6.0.3
certifi==2019.9.11
cffi==1.13.1
cryptography==2.8
Django==2.2.6
future==0.18.2
image==1.5.27
Pillow==6.2.1
pycparser==2.19
pytesseract==0.3.0
python-dotenv==0.10.3
python-telegram-bot==12.2.0
pytz==2019.3
six==1.12.0
sqlparse==0.3.0
tornado==6.0.3
